+++ lib/legion/data/models/identity/identity.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require_relative 'model_helpers'
+
+return unless Legion::Data::Model::Identity::ModelHelpers.table_available?(:portable_identities)
+
+module Legion
+  module Data
+    module Model
+      class Identity
+        class Identity < Sequel::Model(:portable_identities)
+          include ModelHelpers
+
+          many_to_one :principal, class: 'Legion::Data::Model::Identity::Principal'
+          many_to_one :provider, class: 'Legion::Data::Model::Identity::Provider', key: :provider_id
+
+          def self.lookup_columns
+            %i[id uuid provider_identity_key]
+          end
+        end
+      end
+    end
+  end
+end