+++ lib/legion/data/models/runner.rb
@@ -5,8 +5,6 @@ module Legion
   module Data
     module Model
       class Runner < Sequel::Model
-        many_to_one :chain
-        one_to_many :task
         one_to_many :functions
         many_to_one :extension
       end